Update `fitAsync` example in README to describe callback
Thanks for sharing this awesome work!

I made the wrong assumption about how this API worked reading the doc, figuring the callback would return the value.  I think it's because the method uses both a callback and async/await, so I thought it'd help clarify that the callback is for tracking progress and stopping early.